design-projects/micro-services-with-open-telemetry/processor/requirements.txt
fastapi
uvicorn
requests
pydantic
sqlalchemy
psycopg2-binary
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-requests
